+++ test/test-appcd-config.js
@@ -1,3 +1,4 @@
+import fs from 'fs-extra';
 import os from 'os';
 import path from 'path';
 import {
@@ -9,7 +10,7 @@ import {
 const configFile = path.join(os.homedir(), '.appcelerator', 'appcd', 'config.json');
 
 describe('appcd config', function () {
-	this.timeout(60000);
+	this.timeout(120000);
 
 	describe('help', () => {
 		it('should output help as JSON', makeTest(async function () {
@@ -92,15 +93,22 @@ describe('appcd config', function () {
 		'started'
 	];
 
-	const actions = [
+	let getActions = [
 		'get',
 		'ls',
 		'list'
 	];
 
+	let removeActions = [
+		'delete',
+		'rm',
+		'remove',
+		'unset'
+	];
+
 	for (const appcdState of states) {
 		describe(`appcd ${appcdState}`, () => {
-			for (const action of actions) {
+			for (const action of getActions) {
 				describe(action, () => {
 					it(`should ${action} default config as text with banner`, makeTest(async function () {
 						if (appcdState === 'started') {
@@ -360,177 +368,154 @@ describe('appcd config', function () {
 					}
 				});
 			}
+
+			describe('set', () => {
+				it('should error when calling set without a name or value', makeTest(async function () {
+					if (appcdState === 'started') {
+						await this.installNode();
+						await this.startDaemonDebugMode(defaultConfig);
+					}
+
+					const { status, stderr } = this.runAppcdSync([ 'config', 'set' ]);
+					expect(status).to.equal(1);
+					expect(stderr.toString().split('\n\n')[0]).to.equal('Error: Missing the configuration key to set');
+				}));
+
+				it('should error when calling set without a value', makeTest(async function () {
+					if (appcdState === 'started') {
+						await this.installNode();
+						await this.startDaemonDebugMode(defaultConfig);
+					}
+
+					const { status, stderr } = this.runAppcdSync([ 'config', 'set', 'appcd-test' ]);
+					expect(status).to.equal(1);
+					expect(stderr.toString().split('\n\n')[0]).to.equal('Error: Missing the configuration value to set');
+				}));
+
+				it('should set a value with no existing config file', makeTest(async function () {
+					if (appcdState === 'started') {
+						await this.installNode();
+						await this.startDaemonDebugMode(defaultConfig);
+					}
+
+					expect(fs.existsSync(configFile)).to.be.false;
+					const { status, stdout } = this.runAppcdSync([ 'config', 'set', 'appcd-test', 'it works!' ]);
+
+					expect(status).to.equal(0);
+					expect(stdout.toString().split('\n\n')[1].trim()).to.equal('Saved');
+
+					expect(fs.existsSync(configFile)).to.be.true;
+					expect(fs.readJsonSync(configFile)).to.deep.equal({
+						'appcd-test': 'it works!'
+					});
+				}));
+
+				it('should set a value with no existing config file as JSON', makeTest(async function () {
+					if (appcdState === 'started') {
+						await this.installNode();
+						await this.startDaemonDebugMode(defaultConfig);
+					}
+
+					expect(fs.existsSync(configFile)).to.be.false;
+					const { status, stdout } = this.runAppcdSync([ 'config', '--json', 'set', 'appcd-test', 'it works!' ]);
+
+					expect(status).to.equal(0);
+					expect(JSON.parse(stdout)).to.deep.equal({
+						code: 0,
+						result: 'Saved'
+					});
+
+					expect(fs.existsSync(configFile)).to.be.true;
+					expect(fs.readJsonSync(configFile)).to.deep.equal({
+						'appcd-test': 'it works!'
+					});
+				}));
+			});
+
+			for (const action of removeActions) {
+				describe(action, () => {
+					it(`should ${action} config value`, makeTest(async function () {
+						await this.initHomeDir(path.join(__dirname, 'fixtures', 'foo-config'));
+
+						if (appcdState === 'started') {
+							await this.installNode();
+							await this.startDaemonDebugMode(defaultConfig);
+						}
+
+						const { status, stdout } = this.runAppcdSync([ 'config', '--json', action, 'foo' ], {}, defaultConfig);
+
+						expect(status).to.equal(0);
+
+						expect(JSON.parse(stdout)).to.deep.equal({
+							code: 0,
+							result: 'Saved'
+						});
+
+						expect(fs.existsSync(configFile)).to.be.true;
+						expect(fs.readJsonSync(configFile)).to.deep.equal({});
+					}));
+				});
+			}
+
+			describe.only('arrays', () => {
+				it('should push to an array config value', makeTest(async function () {
+					await this.initHomeDir(path.join(__dirname, 'fixtures', 'array-config'));
+
+					if (appcdState === 'started') {
+						await this.installNode();
+						await this.startDaemonDebugMode(defaultConfig);
+					}
+
+					const { status, stdout } = this.runAppcdSync([ 'config', '--json', 'push', 'foo', 'bar' ]);
+					expect(status).to.equal(0);
+					expect(JSON.parse(stdout).result).to.deep.equal([ 'a', 'b', 'c', 'bar' ]);
+					expect(fs.readJsonSync(configFile)).to.deep.equal({
+						foo: [ 'a', 'b', 'c', 'bar' ]
+					});
+				}));
+
+				it('should pop from an array config value', makeTest(async function () {
+					await this.initHomeDir(path.join(__dirname, 'fixtures', 'array-config'));
+
+					if (appcdState === 'started') {
+						await this.installNode();
+						await this.startDaemonDebugMode(defaultConfig);
+					}
+
+					const { status, stdout } = this.runAppcdSync([ 'config', '--json', 'pop', 'foo' ]);
+					expect(status).to.equal(0);
+					expect(JSON.parse(stdout).result).to.equal('c');
+					expect(fs.readJsonSync(configFile)).to.deep.equal({ foo: [ 'a', 'b' ] });
+				}));
+
+				it('should shift an array config value', makeTest(async function () {
+					await this.initHomeDir(path.join(__dirname, 'fixtures', 'array-config'));
+
+					if (appcdState === 'started') {
+						await this.installNode();
+						await this.startDaemonDebugMode(defaultConfig);
+					}
+
+					const { status, stdout } = this.runAppcdSync([ 'config', '--json', 'shift', 'foo' ]);
+					expect(status).to.equal(0);
+					expect(JSON.parse(stdout).result).to.equal('a');
+					expect(fs.readJsonSync(configFile)).to.deep.equal({ foo: [ 'b', 'c' ] });
+				}));
+
+				it('should unshift an array config value', makeTest(async function () {
+					await this.initHomeDir(path.join(__dirname, 'fixtures', 'array-config'));
+
+					if (appcdState === 'started') {
+						await this.installNode();
+						await this.startDaemonDebugMode(defaultConfig);
+					}
+
+					const { status, stdout } = this.runAppcdSync([ 'config', '--json', 'unshift', 'foo', 'bar' ]);
+					expect(status).to.equal(0);
+					expect(JSON.parse(stdout).result).to.deep.equal([ 'bar', 'a', 'b', 'c' ]);
+					expect(fs.readJsonSync(configFile)).to.deep.equal({ foo: [ 'bar', 'a', 'b', 'c' ] });
+				}));
+			});
 		});
 	}
-
-	describe.skip('set', () => {
-		// it('should set a value with no existing config file', async function () {
-		// 	return runSync({
-		// 		args: [ 'config', '--json', 'set', 'foo', 'bar' ],
-		// 		then({ status, stdout }) {
-		// 			expect(status).to.equal(0);
-		// 			expect(JSON.parse(stdout)).to.deep.equal({
-		// 				code: 0,
-		// 				result: 'Saved'
-		// 			});
-		// 			expect(fs.readJSONSync(configFile)).to.deep.equal({ foo: 'bar' });
-		// 		}
-		// 	});
-		// });
-	});
 });
-
-/*
-const { cleanConfig, preCheck, readConfig, restoreConfigFile, runJSONCommand, writeConfig } = require('./utils');
-let backupFile;
-
-describe('amplify config integration tests', function () {
-	this.timeout(5000);
-
-	before(function () {
-		backupFile = preCheck();
-	});
-
-	beforeEach(function () {
-		cleanConfig();
-	});
-
-	after(function () {
-		if (backupFile) {
-			restoreConfigFile(backupFile);
-		}
-	});
-
-	[ 'get', 'ls', 'list'].forEach( function (getCommand) {
-		it(`config can list a specific value with ${getCommand}`, async function () {
-			writeConfig({
-				foo: 'bar'
-			});
-
-			const getCmd =  await runJSONCommand([ 'config', getCommand, 'foo' ]);
-			expect(getCmd.code).to.equal(0);
-			expect(getCmd.stdout.result).to.equal('bar');
-		});
-
-		it(`config can list entire config with ${getCommand}`, async function () {
-			writeConfig({
-				foo: 'bar',
-				bar: 'foo'
-			});
-
-			const getCmd =  await runJSONCommand([ 'config', getCommand ]);
-			expect(getCmd.code).to.equal(0);
-			expect(getCmd.stdout.code).to.equal(0);
-			expect(getCmd.stdout.result).to.deep.equal({ bar: 'foo', foo: 'bar' });
-		});
-	})
-
-	it('config can list entire config', async function () {
-		writeConfig({
-			foo: 'bar',
-			bar: 'foo'
-		});
-
-		const getCmd =  await runJSONCommand([ 'config', 'get' ]);
-		expect(getCmd.code).to.equal(0);
-		expect(getCmd.stdout.code).to.equal(0);
-		expect(getCmd.stdout.result).to.deep.equal({ bar: 'foo', foo: 'bar' });
-	});
-
-	[ 'delete', 'rm', 'remove', 'unset' ].forEach(function(removalCommand) {
-
-		it(`config can delete values with ${removalCommand}`, async function () {
-			writeConfig({
-				foo: 'bar'
-			});
-
-			const deleteCmd = await runJSONCommand([ 'config', removalCommand, 'foo' ]);
-			expect(deleteCmd.code).to.equal(0);
-			expect(deleteCmd.stdout.code).to.equal(0);
-			expect(deleteCmd.stdout.result).to.equal('Saved');
-
-			const getCmd =  await runJSONCommand([ 'config', 'get', 'foo' ]);
-			expect(getCmd.code).to.equal(6);
-			expect(getCmd.code).to.equal(6);
-			expect(getCmd.stdout.result).to.equal('Not Found: foo');
-		});
-
-	});
-
-	it('config can push to arrays', async function () {
-		writeConfig({
-			foo: [ 'avalue' ]
-		});
-
-		const pushCmd = await runJSONCommand([ 'config', 'push', 'foo', 'bar' ]);
-		expect(pushCmd.code).to.equal(0);
-		expect(pushCmd.code).to.equal(0);
-		expect(pushCmd.stdout.result).to.deep.equal([ 'avalue', 'bar' ]);
-
-		const config = readConfig();
-		expect(config).to.deep.equal({ foo: [ 'avalue', 'bar' ] });
-
-		const invalidShiftCmd = await runJSONCommand([ 'config', 'push', 'bar', 'foo' ]);
-		expect(invalidShiftCmd.code).to.equal(0);
-		expect(invalidShiftCmd.stdout.code).to.equal(0);
-		expect(invalidShiftCmd.stdout.result).to.deep.equal([ 'foo' ]);
-	});
-
-	it('config can pop values from arrays', async function () {
-		writeConfig({
-			foo: [ 'avalue', 'poppedval' ]
-		});
-
-		const popCmd = await runJSONCommand([ 'config', 'pop', 'foo' ]);
-		expect(popCmd.code).to.equal(0);
-		expect(popCmd.code).to.equal(0);
-		expect(popCmd.stdout.result).to.equal('poppedval');
-
-		const config = readConfig();
-		expect(config).to.deep.equal({ foo: [ 'avalue' ] });
-
-		const invalidPopCmd = await runJSONCommand([ 'config', 'pop', 'bar' ]);
-		expect(invalidPopCmd.code).to.equal(6);
-		expect(invalidPopCmd.stdout.code).to.equal(6);
-		expect(invalidPopCmd.stdout.result).to.equal('Not Found: bar');
-	});
-
-	it('config can shift values from arrays', async function () {
-		writeConfig({
-			foo: [ 'shiftedval', 'bar' ]
-		});
-
-		const shiftCmd = await runJSONCommand([ 'config', 'shift', 'foo' ]);
-		expect(shiftCmd.code).to.equal(0);
-		expect(shiftCmd.code).to.equal(0);
-		expect(shiftCmd.stdout.result).to.equal('shiftedval');
-
-		const config = readConfig();
-		expect(config).to.deep.equal({ foo: [ 'bar' ] });
-
-		const invalidShiftCmd = await runJSONCommand([ 'config', 'shift', 'bar' ]);
-		expect(invalidShiftCmd.code).to.equal(6);
-		expect(invalidShiftCmd.stdout.code).to.equal(6);
-		expect(invalidShiftCmd.stdout.result).to.equal('Not Found: bar');
-	});
-
-	it('config can unshift values to an array', async function () {
-		writeConfig({
-			foo: [ 'bar' ]
-		});
-
-		const pushCmd = await runJSONCommand([ 'config', 'unshift', 'foo', 'unshiftedval' ]);
-		expect(pushCmd.code).to.equal(0);
-		expect(pushCmd.code).to.equal(0);
-		expect(pushCmd.stdout.result).to.deep.equal([ 'unshiftedval', 'bar']);
-
-		const config = readConfig();
-		expect(config).to.deep.equal({ foo: [ 'unshiftedval', 'bar' ] });
-
-		const invalidShiftCmd = await runJSONCommand([ 'config', 'unshift', 'bar', 'foo' ]);
-		expect(invalidShiftCmd.code).to.equal(0);
-		expect(invalidShiftCmd.stdout.code).to.equal(0);
-		expect(invalidShiftCmd.stdout.result).to.deep.equal([ 'foo' ]);
-	});
-});
-*/