+++ test/test-appcd-start.js
@@ -5,7 +5,7 @@ import {
 } from './common';
 
 describe('appcd start', function () {
-	this.timeout(60000);
+	this.timeout(120000);
 
 	describe('help', () => {
 		it('should output help as JSON', makeTest(async function () {