+++ test/test-plugin-jdk.js
@@ -16,7 +16,7 @@ try {
 }
 
 describe('plugin jdk', function () {
-	this.timeout(60000);
+	this.timeout(120000);
 
 	_it('should register the jdk plugin', makeTest(async function () {
 		this.symlinkPlugin('jdk', pluginVersion);